build on xenial

+++ src/value.c
@@ -203,7 +203,7 @@ int SIValue_ToString(SIValue v, char *buf, size_t len) {
     bytes_written = snprintf(buf, len, "%d", v.intval);
     break;
   case T_INT64:
-    bytes_written = snprintf(buf, len, "%lld", v.longval);
+    bytes_written = snprintf(buf, len, "%lld", (long long)v.longval);
     break;
   case T_UINT:
     bytes_written = snprintf(buf, len, "%zd", v.uintval);
@@ -274,7 +274,7 @@ int SI_LongVal_Cast(SIValue *v, SIType type) {
     break;
   case T_STRING: {
     char *buf = malloc(21);
-    snprintf(buf, 21, "%lld", v->longval);
+    snprintf(buf, 21, "%lld", (long long)v->longval);
     v->stringval = SI_StringValC(buf).stringval;
     break;
   }
